fix: revert unrelated combinators test changes from telemetry PR
