+++ combinators/shared/src/test/scala-3/zio/blocks/combinators/TypeInferenceSpec.scala
@@ -16,9 +16,8 @@
 
 package zio.blocks.combinators
 
-import scala.compiletime.testing.typeCheckErrors
-
 import zio.test._
+import scala.compiletime.testing.typeCheckErrors
 
 object TypeInferenceSpec extends ZIOSpecDefault {
